찾기 페이지에 영상이 보여지도록 수정
