--- src/error.py
@@ -1,6 +1,0 @@
-class QRCodeError(Exception):
-    pass
-
-
-class VideoError(Exception):
-    pass
